requirements.txt
python-telegram-bot[job-queue]==21.3
yt_dlp
